unify unnecessary files
