+++ database/engine/engine.go
@@ -12,3 +12,14 @@ type Transaction interface {
 	Commit() error
 	Discard()
 }
+
+type Snapshot interface {
+	Get(key []byte) ([]byte, error)
+	Has(key []byte) (bool, error)
+	NewIterator(start, end []byte) (Iterator, error)
+	Releaser
+}
+
+type Releaser interface {
+	Release()
+}
